+++ Gopkg.toml
@@ -57,6 +57,10 @@
   name = "github.com/tyler-smith/go-bip39"
   branch = "master"
 
+[[constraint]]
+  name = "github.com/zondax/ledger-goclient"
+  branch = "master"
+
 [prune]
   go-tests = true
   unused-packages = true
